Update puma config, disconnect from databases before forking (#23)
Add a `config/puma.rb`, matching what we’ll generate for new Hanami 2.1 apps.

Update the persistence provider to disconnect from the database in its `stop` hook. This fits nicely with the `Hanami.shutdown` we run in Puma's `before_fork` hook, and in doing this, we apply Sequel’s fork safety advice of disconnecting before forking.