Update dev setup: use Docker Compose and Mise
Switch from just expecting a postgres to be running _somehow_ to running it ourself in Docker Compose, for a more self-contained dev setup.

While here, make it clear we prefer Mise for Ruby and Node versions. Switch to mise’s native config file and include mise (as well as docker compose) steps in README.